+++ tests/Integration/Schema/Directives/AllDirectiveTest.php
@@ -5,7 +5,7 @@ namespace Tests\Integration\Schema\Directives;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection as EloquentCollection;
 use Illuminate\Database\Eloquent\Relations\Relation;
-use Mockery;
+use Laravel\Scout\Builder as ScoutBuilder;
 use Tests\DBTestCase;
 use Tests\TestsScoutEngine;
 use Tests\Utils\Models\Post;
@@ -14,6 +14,8 @@ use Tests\Utils\Models\User;
 final class AllDirectiveTest extends DBTestCase
 {
     use TestsScoutEngine;
+
+    public const LIMIT_FROM_CUSTOM_SCOUT_BUILDER = 321;
 
     public function testGetAllModelsAsRootField(): void
     {
@@ -258,82 +260,15 @@ final class AllDirectiveTest extends DBTestCase
     {
         $this->setUpScoutEngine();
 
-        /** @var \Tests\Utils\Models\Post $postA */
-        $postA = factory(Post::class)->create([
-            'title' => 'great title',
-        ]);
-        /** @var \Tests\Utils\Models\Post $postB */
-        $postB = factory(Post::class)->create([
-            'title' => 'Really great title',
-        ]);
-        factory(Post::class)->create([
-            'title' => 'bad title',
-        ]);
+        $post = factory(Post::class)->create();
+        assert($post instanceof Post);
 
         $this->engine->shouldReceive('map')
-            ->andReturn(
-                new EloquentCollection([$postA, $postB])
-            )
-            ->once();
-
-        $this->schema = /** @lang GraphQL */ <<<GRAPHQL
-        type Post {
-            id: ID!
-        }
-
-        type Query {
-            posts: [Post!]! @all(builder: "{$this->qualifyTestResolver('builderForScoutBuilder')}")
-        }
-GRAPHQL;
-
-        $this->graphQL(/** @lang GraphQL */ '
-        {
-            posts {
-                id
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'posts' => [
-                    [
-                        'id' => '1',
-                    ],
-                    [
-                        'id' => '2',
-                    ],
-                ],
-            ],
-        ]);
-    }
-
-    public function testSpecifyCustomBuilderForScoutBuilderWithScoutDirective(): void
-    {
-        $this->setUpScoutEngine();
-
-        /** @var \Mockery\MockInterface&\Laravel\Scout\Builder $builder */
-        $builder = Mockery::mock(Post::search())->makePartial();
-        app()->bind(\Laravel\Scout\Builder::class, function () use ($builder) {
-            return $builder;
-        });
-
-        /** @var \Tests\Utils\Models\Post $postA */
-        $postA = factory(Post::class)->create([
-            'title' => 'great title',
-            'task_id' => 1,
-        ]);
-        factory(Post::class)->create([
-            'title' => 'Really great title',
-            'task_id' => 2,
-        ]);
-        factory(Post::class)->create([
-            'title' => 'bad title',
-            'task_id' => 3,
-        ]);
-
-        $this->engine->shouldReceive('map')
-            ->andReturn(
-                new EloquentCollection([$postA])
-            )
+            ->withArgs(function (ScoutBuilder $builder) use ($post): bool {
+                return $builder->wheres === ['id' => "$post->id"]
+                    && self::LIMIT_FROM_CUSTOM_SCOUT_BUILDER === $builder->limit;
+            })
+            ->andReturn(new EloquentCollection([$post]))
             ->once();
 
         $this->schema = /** @lang GraphQL */ <<<GRAPHQL
@@ -343,34 +278,28 @@ GRAPHQL;
 
         type Query {
             posts(
-                task: ID! @eq(key: "task_id")
-            ): [Post!]!
-                @all(builder: "{$this->qualifyTestResolver('builderForScoutBuilder')}")
+                id: ID! @eq
+            ): [Post!]! @all(builder: "{$this->qualifyTestResolver('builderForScoutBuilder')}")
         }
 GRAPHQL;
 
         $this->graphQL(/** @lang GraphQL */ '
-        {
-            posts(task: "1") {
+        query ($id: ID!) {
+            posts(id: $id) {
                 id
             }
         }
-        ')->assertJson([
+        ', [
+            'id' => $post->id,
+        ])->assertJson([
             'data' => [
                 'posts' => [
                     [
-                        'id' => '1',
+                        'id' => "$post->id",
                     ],
                 ],
             ],
         ]);
-
-        // Ensure `@eq` directive has been applied on scout builder instance
-        $builder->shouldHaveReceived('where')
-            ->with(
-                'task_id',
-                '1'
-            );
     }
 
     public function builder(): Builder
@@ -383,8 +312,9 @@ GRAPHQL;
         return $parent->posts()->orderBy('id', 'DESC');
     }
 
-    public function builderForScoutBuilder(): \Laravel\Scout\Builder
+    public function builderForScoutBuilder(): ScoutBuilder
     {
-        return Post::search('great title');
+        return Post::search('great title')
+            ->take(self::LIMIT_FROM_CUSTOM_SCOUT_BUILDER);
     }
 }