+++ tests/TestsScoutEngine.php
@@ -22,12 +22,12 @@ trait TestsScoutEngine
     public function setUpScoutEngine(): void
     {
         $this->engineManager = Mockery::mock(EngineManager::class);
-        $this->engine = Mockery::mock(NullEngine::class)
-            ->makePartial();
-
         $this->app->singleton(EngineManager::class, function (): MockInterface {
             return $this->engineManager;
         });
+
+        $this->engine = Mockery::mock(NullEngine::class)
+            ->makePartial();
 
         $this->engineManager
             ->shouldReceive('engine')